separate bin in two files: `node-core-test` and `node--test`
