+++ README.md
@@ -14,9 +14,6 @@ node core.
 Differences from the core implementation:
 
 - Doesn't hide its own stack frames
-- Internally uses `._private` property names instead of `#private` fields,
-  for compatibility
-- Uses `String` instead of `Symbol`, for compatibility
 
 ## Docs
 
@@ -163,14 +160,14 @@ test('skip() method with message', t => {
 
 ### `only` tests
 
-If Node.js is started with the `--test-only` command-line option, it is
+If `node-core-test` is started with the `--test-only` command-line option, it is
 possible to skip all top level tests except for a selected subset by passing
 the `only` option to the tests that should be run. When a test with the `only`
 option set is run, all subtests are also run. The test context's `runOnly()`
 method can be used to implement the same behavior at the subtest level.
 
 ```js
-// Assume Node.js is run with the --test-only command-line option.
+// Assume node-core-test is run with the --test-only command-line option.
 // The 'only' option is set, so this test is run.
 test('this test is run', { only: true }, async t => {
   // Within this test, all subtests are run by default.
@@ -237,7 +234,9 @@ test('a test that creates asynchronous activity', t => {
 The Node.js test runner can be invoked from the command line:
 
 ```bash
-node-core-test
+node-core-test --test
+# or use the shortcut version:
+node--test
 ```
 
 By default, Node.js will recursively search the current directory for
@@ -250,7 +249,8 @@ Alternatively, one or more paths can be provided as the final argument(s) to
 the Node.js command, as shown below.
 
 ```bash
-node-core-test test1.js test2.mjs custom_test_dir/
+node-core-test --test test1.js test2.mjs custom_test_dir/
+node--test test1.js test2.mjs custom_test_dir/
 ```
 
 In this example, the test runner will execute the files `test1.js` and
