+++ test/fixtures/node-core-test/esm.test.mjs
